@@ -4,42 +4,32 @@ namespace h4kuna\Queue;
 
 use h4kuna\Memoize\MemoryStorage;
 use h4kuna\Queue\Exceptions;
+use SysvMessageQueue;
 
 final class Queue
 {
 	use MemoryStorage;
 
 	public const MAX_MESSAGE_SIZE = 256; // bytes, max is by system, observed 8192
-	public const INFO_UID = 'msg_perm.uid';
-	public const INFO_GID = 'msg_perm.gid';
-	public const INFO_MODE = 'msg_perm.mode';
+	public const INFO_SETUP_UID = 'msg_perm.uid';
+	public const INFO_SETUP_GID = 'msg_perm.gid';
+	public const INFO_SETUP_MODE = 'msg_perm.mode';
 	public const INFO_CREATE_TIME = 'msg_ctime';
 	public const INFO_SEND_TIME = 'msg_stime';
 	public const INFO_RECEIVE_TIME = 'msg_rtime';
 	public const INFO_COUNT = 'msg_qnum';
-	public const INFO_BYTES = 'msg_qbytes';
+	public const INFO_SETUP_BYTES = 'msg_qbytes';
 	public const INFO_LAST_SEND_PID = 'msg_lspid';
 	public const INFO_LAST_RECEIVE_PID = 'msg_lrpid';
 
-	/** @var string */
-	private $name;
 
-	/** @var int */
-	private $key;
-
-	/** @var int */
-	private $permission;
-
-	/** @var int */
-	private $maxMessageSize;
-
-
-	public function __construct(string $name, int $key, int $permission, int $maxMessageSize = self::MAX_MESSAGE_SIZE)
+	public function __construct(
+		private string $name,
+		private int $key,
+		private int $permission,
+		private int $maxMessageSize = self::MAX_MESSAGE_SIZE,
+	)
 	{
-		$this->name = $name;
-		$this->key = $key;
-		$this->permission = $permission;
-		$this->maxMessageSize = $maxMessageSize;
 	}
 
 
@@ -51,14 +41,15 @@ final class Queue
 
 	public function fullname(): string
 	{
-		if (((int) $this->name) === $this->key) {
+		if (intval($this->name) === $this->key) {
 			return $this->name;
 		}
+
 		return "$this->name.$this->key";
 	}
 
 
-	public function resource()
+	public function resource(): SysvMessageQueue
 	{
 		return $this->createResource();
 	}
@@ -70,16 +61,20 @@ final class Queue
 	}
 
 
+	/**
+	 * @param array<self::INFO_SETUP_*, int> $data
+	 * @return bool
+	 */
 	public function setup(array $data): bool
 	{
-		$structure = [self::INFO_UID, self::INFO_GID, self::INFO_MODE, self::INFO_BYTES];
+		$structure = [self::INFO_SETUP_UID, self::INFO_SETUP_GID, self::INFO_SETUP_MODE, self::INFO_SETUP_BYTES];
 
 		return msg_set_queue($this->resource(), array_intersect_key($data, array_fill_keys($structure, true)));
 	}
 
 
 	/**
-	 * @return array<string, mixed>
+	 * @return array<self::INFO_*, int>
 	 * @throws Exceptions\QueueInfoIsUnavailableException
 	 */
 	public function info(): array
@@ -101,20 +96,20 @@ final class Queue
 	{
 		$info = $this->info();
 		$extends = [
-			self::INFO_MODE => Linux::permissionInToText($info[self::INFO_MODE]),
+			self::INFO_SETUP_MODE => Linux::permissionInToText($info[self::INFO_SETUP_MODE]),
 			self::INFO_CREATE_TIME => self::createDateTime($info[self::INFO_CREATE_TIME]),
 			self::INFO_SEND_TIME => self::createDateTime($info[self::INFO_SEND_TIME]),
 			self::INFO_RECEIVE_TIME => self::createDateTime($info[self::INFO_RECEIVE_TIME]),
-			self::INFO_BYTES => $info[self::INFO_BYTES],
+			self::INFO_SETUP_BYTES => $info[self::INFO_SETUP_BYTES],
 			self::INFO_COUNT => $info[self::INFO_COUNT],
 			self::INFO_LAST_RECEIVE_PID => $info[self::INFO_LAST_RECEIVE_PID],
 			self::INFO_LAST_SEND_PID => $info[self::INFO_LAST_SEND_PID],
 		];
 
 		[
-			'user' => $extends[self::INFO_UID],
-			'group' => $extends[self::INFO_GID],
-		] = Linux::userGroupToText($info[self::INFO_UID], $info[self::INFO_GID]);
+			'user' => $extends[self::INFO_SETUP_UID],
+			'group' => $extends[self::INFO_SETUP_GID],
+		] = Linux::userGroupToText($info[self::INFO_SETUP_UID], $info[self::INFO_SETUP_GID]);
 
 		return $extends;
 	}
@@ -154,27 +149,26 @@ final class Queue
 	}
 
 
-	/**
-	 * @return resource
-	 */
-	private function createResource()
+	private function createResource(): SysvMessageQueue
 	{
 		$exists = msg_queue_exists($this->key);
 		$queue = msg_get_queue($this->key, $this->permission);
+
+		if ($queue === false) {
+			throw new Exceptions\CreateQueueException(sprintf('Queue "%s" failed to create.', $this->fullname()));
+		}
 
 		if ($exists && ($perm = $this->queuePermission($queue)) !== $this->permission) {
 			throw new Exceptions\CreateQueueException(sprintf('Queue "%s" already exists with permissions "%s" and you require "%s". %s',
 				$this->fullname(), $perm, $this->permission, $this->helpHowRemove($perm)));
 		}
 
+
 		return $queue;
 	}
 
 
-	/**
-	 * @param resource $queue
-	 */
-	private function queuePermission($queue): int
+	private function queuePermission(SysvMessageQueue $queue): int
 	{
 		$stats = msg_stat_queue($queue);
 		if (!is_array($stats)) {
@@ -182,17 +176,17 @@ final class Queue
 				$this->helpHowRemove($this->permission)));
 		}
 
-		return $stats[self::INFO_MODE];
+		return $stats[self::INFO_SETUP_MODE];
 	}
 
 
-	private static function createDateTime(int $timestamp): ?\DateTime
+	private static function createDateTime(int $timestamp): ?\DateTimeImmutable
 	{
 		if ($timestamp === 0) {
 			return null;
 		}
 
-		return (new \DateTime("@$timestamp"))->setTimezone(new \DateTimeZone(date_default_timezone_get()));
+		return new \DateTimeImmutable("@$timestamp");
 	}
 
 
