@@ -2,8 +2,6 @@
 
 require __DIR__ . '/../vendor/autoload.php';
 
-Salamium\Testinium\File::setRoot(__DIR__ . '/data/tests');
-
 date_default_timezone_set('Europe/Prague');
 Tester\Environment::setup();
-\Tracy\Debugger::enable(false, __DIR__ . '/temp');
+Tracy\Debugger::enable(false, __DIR__ . '/temp');
